+++ trainings/2015-10-13-2012-Sichuan-Province-Programming-Contest/B.cpp
@@ -8,7 +8,7 @@ int k[N], a[N];
 
 typedef long long LL;
 
-LL get_num(int x) {
+long long get_num(int x) {
     return (1LL * x * x + x) / 2;
 }
 
@@ -25,7 +25,7 @@ void work() {
             k[i] = a[i] >> pos & 1;
         }
         num[0] = num[1] = 0;
-        LL ans_and = 0, ans_or = 0, ans_xor = 0;
+        long long ans_and = 0, ans_or = 0, ans_xor = 0;
         for (int i = 1, j; i <= n; i = j) {
             if (k[i] == 0) {
                 for (j = i; j <= n && k[j] == 0; j++);
@@ -43,7 +43,7 @@ void work() {
             num[now]++;
         }
         //	printf("%lld %lld %lld\n", ans_and, ans_or, ans_xor);
-        LL t = get_num(n);
+        long long t = get_num(n);
         andd += (1. * ans_and / t) * (1 << pos);
         orr += (1. - (1. * ans_or / t)) * (1 << pos);
         xorr += (1. * ans_xor / t) * (1 << pos);
